birdseye-backend/requirements.txt
absl-py==2.3.1
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
fastapi==0.115.6
filelock==3.20.3
filetype==1.2.0
flatbuffers==25.9.23
fonttools==4.61.1
fsspec==2026.1.0
idna==3.7
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
mediapipe==0.10.9
mpmath==1.3.0
networkx==3.6.1
numpy==2.2.6
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opencv-python-headless==4.10.0.84
packaging==25.0
pi_heif==1.1.1
pillow==12.0.0
pillow-avif-plugin==1.5.2
polars==1.37.1
polars-runtime-32==1.37.1
protobuf==3.20.3
psutil==7.2.1
pycparser==2.23
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
roboflow==1.2.12
scipy==1.17.0
six==1.17.0
sounddevice==0.5.3
sympy==1.13.1
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
typing_extensions==4.15.0
ultralytics==8.4.5
ultralytics-thop==2.0.18
urllib3==2.6.3
uvicorn==0.34.0